Make booking-audit integration test utils reusable
